Auslagern der Cards
Auslagern der Cards in favoriten, myRecipesMenu und recipesMenu
Überarbeitung der recipeCard
erstellen der recipeCard.scss
