+++ src/components/MyRecipesMenu.jsx
@@ -19,7 +19,7 @@ const MyRecipesMenu = () => {
                 allRecipes.push (
                     <React.Fragment key={product.id}>
                         <div class="oneRecipe p-m-1 p-p-1 p-m-lg-3 p-b-lg-3"> 
-                        <Card className="p-mb-2" title={product.name} subTitle={product.categories} style={{ width: '25em', height:'30em' }} footer={product.averagetimeinminutes + " Minuten"} header={<div><img alt="pictureRecipe" src={product.imageurl}></img></div>}></Card>
+                        <RecipeCard cardProduct = {product}></RecipeCard>
                         </div>
                     </React.Fragment>
                 );
