+++ src/components/favoriten.jsx
@@ -21,10 +21,7 @@ const Favoriten = () => {
                 allRecipes.push (
                     <React.Fragment key={product.id}>
                         <div class="oneRecipe p-m-1 p-p-1 p-m-lg-3 p-b-lg-3"> 
-                        <Card className="p-mb-2" title={product.name} subTitle={product.categories} style={{ width: '25em', height:'30em' }} footer={product.averagetimeinminutes + " Minuten"} header={<div><img alt="pictureRecipe" src={product.imageurl}></img></div>}>
-                        <div className = "likeButton"><Button icon="pi pi-heart" className="p-button-rounded p-button-help p-button-text" 
-                            /*onClick={setfavorite}*/ /></div>
-                        </Card>
+                        <RecipeCard cardProduct = {product}></RecipeCard>
                         </div>
                     </React.Fragment>
                 );
