+++ src/components/recipeCard.jsx
@@ -2,38 +2,47 @@ import React, { useEffect, useState } from "react";
 import { Card } from 'primereact/card';
 import axios from "axios";
 import { Button } from 'primereact/button';
-
-const RecipeCard = () => {
-    const [products, setProducts] = useState([{id: 1, name: "SUPERyummy", imageurl: "https://cdn.pixabay.com/photo/2017/05/25/21/22/pizza-2344399_960_720.jpg", categories: "[Hauptspeise, Vegetarisch, Vegan, Backen]" },{id: 1, name: "SUPERyummy", imageurl: "https://cdn.pixabay.com/photo/2017/05/25/21/22/pizza-2344399_960_720.jpg", categories: "[Hauptspeise, Vegetarisch, Vegan, Backen]" },{id: 1, name: "SUPERyummy", imageurl: "https://cdn.pixabay.com/photo/2017/05/25/21/22/pizza-2344399_960_720.jpg", categories: "[Hauptspeise, Vegetarisch, Vegan, Backen]" }]);
-    const [currentPosition, setCurrentPosition] = useState(3);
-    const [final, setFinal] = useState([]);
+import { Chip } from 'primereact/chip';
 
 
 
+const RecipeCard = ({cardProduct}) => {
+    const [loginStatus, setLogingStatus] = useState([]);
 
-    const cardheader = (x) => (
-        <img src={x.imageurl} alt="recipe" width="500"/>
-    );  
+    const   setfavorite = async () => {
 
-    const cardfooter = (
-        <span>
-            <Button label="Favorit" icon="pi pi-heart" />
-        </span>
-    );
+         const res = await axios({
+             method: "post",
+             url: "https://ratatouilleexpress.retch.duckdns.org/api/recipes/like",
+             data: {
+                 "accountId": window.$id,
+                 "recipeId": cardProduct.id 
+                 
+             }
+           }).catch(error => {
+             return { error: error };
+           });
+        }
 
-    const postSomeRecipes = () =>{
-        let x = [];
-        for (let element of products) {
-          x.push(<Card title={element.name} subTitle={element.categories} style={{ width: '25em' }} footer={cardfooter} header={cardheader(element)}/>);  
-        };
-        setFinal(x);
-    };
+    const checkUserStatus = () =>{
+        if(window.$id === undefined && window.$id !== cardProduct.creatorId)
+        {
+            <p>nicht von mir</p>
+        }
+        else if(window.$id === cardProduct.creatorId){
+            <p>von mir</p>
+        }
+    }
+
+    return(
+    checkUserStatus(),
+    console.log(cardProduct.creatorId),
     
-
-    return (
-        <div class="recipecard"> 
-        {final}
-        </div>
-    );
+    <Card className="p-mb-2" title={cardProduct.name} subTitle={<Chip label={cardProduct.categories} className="p-mr-2 p-mb-2 custom-chip" />} style={{ width: '25em', height:'30em'}} footer={cardProduct.averagetimeinminutes + " Minuten"} header={<div><img alt="pictureRecipe" src={cardProduct.imageurl}></img></div>}>
+        <div onClick={setfavorite} className = "likeButton"><Button icon="pi pi-heart" className="p-button-rounded p-button-help p-button-text"/></div>
+    </Card>
+    
+    )
+    
 }
 export default RecipeCard;
